Fix errors when running the flatpak
No more errors!

+++ flatpak/run.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
-mkdir game
+if [ ! -d ./game ]; then
+	mkdir game
+fi
 cd game
-cp /app/share/forever/help.txt ..
+if [ ! -f ../help.txt ]; then
+	cp /app/share/forever/help.txt ..
+fi
 s1fs2a
